All thats left to adding periodic metric calculations is freeing up memory and calculating the actual jacobian

+++ src/AbstractSingleBlockMesh.hpp
@@ -16,7 +16,7 @@ class AbstractSingleBlockMesh{
 	double *J31, *J32, *J33;
 
 	double max_xi, max_eta, max_zta;
-	double Nx, Ny, Nz;
+	int Nx, Ny, Nz;
 
 	Derivatives *derivX, *derivY, *derivZ;
